+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/containermenu/storage/block/ItemStorageBlockContainerMenu.java
@@ -29,7 +29,8 @@ public class ItemStorageBlockContainerMenu extends AbstractStorageBlockContainer
             PlatformApi.INSTANCE.getResourceTypeRegistry(),
             playerInventory.player,
             buf,
-            ItemResourceType.INSTANCE);
+            ItemResourceType.INSTANCE
+        );
     }
 
     public ItemStorageBlockContainerMenu(final int syncId,