+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/containermenu/transfer/TransferDestination.java
@@ -1,0 +1,10 @@
+package com.refinedmods.refinedstorage2.platform.common.containermenu.transfer;
+
+import javax.annotation.Nullable;
+
+import net.minecraft.world.item.ItemStack;
+
+public interface TransferDestination {
+    @Nullable
+    ItemStack transfer(ItemStack stack);
+}